Adding iOS and Android look and feel for AlertDialogs  with  mohamed Adel

+++ lib/widgets/activity_item.dart
@@ -1,5 +1,6 @@
-import 'package:BorhanAdmin/providers/auth.dart';
+import 'dart:io' show Platform;
 import 'package:BorhanAdmin/providers/organizations_provider.dart';
+import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import '../screens/add_activity.dart';
 import 'package:provider/provider.dart';
@@ -9,7 +10,7 @@ class ActivityItem extends StatefulWidget {
   final String name;
   final String id;
   final String image;
-  final orgLocalId ;
+  final orgLocalId;
 
   ActivityItem({this.name, this.id, this.image, this.orgLocalId});
 
@@ -56,27 +57,80 @@ class _ActivityItemState extends State<ActivityItem> {
                 Navigator.push(
                     context,
                     MaterialPageRoute(
-                        builder: (context) =>
-                            AddActivity(orgLocalId: widget.orgLocalId,actId: widget.id,)));
+                        builder: (context) => AddActivity(
+                              orgLocalId: widget.orgLocalId,
+                              actId: widget.id,
+                            )));
               },
               color: Theme.of(context).primaryColor,
             ),
             IconButton(
               icon: Icon(Icons.delete),
               onPressed: () {
-                Provider.of<Activities>(context, listen: false)
-                    .deleteActivity(widget.id, orgId);
+                _showDialog();
               },
               color: Theme.of(context).errorColor,
             ),
           ],
         ),
       ),
-
       leading: CircleAvatar(
         radius: 25,
         backgroundImage: NetworkImage(widget.image),
       ),
+    );
+  }
+
+  _showDialog() {
+    showDialog(
+      context: context,
+      builder: (ctx) => (Platform.isAndroid)
+          ? AlertDialog(
+              title: Text('حذف نشاط'),
+              content: Text('هل تريد حذف النشاط؟'),
+              actions: <Widget>[
+                FlatButton(
+                  child: Text('الغاء'),
+                  onPressed: () {
+                    Navigator.of(ctx).pop();
+                  },
+                ),
+                FlatButton(
+                  child: Text('نعم',
+                    style: TextStyle(
+                        color: Colors.red, fontWeight: FontWeight.bold),),
+                  onPressed: () {
+                    Provider.of<Activities>(context, listen: false)
+                        .deleteActivity(widget.id, orgId);
+                    Navigator.of(ctx).pop();
+                  },
+                ),
+              ],
+            )
+          : CupertinoAlertDialog(
+              title: Text('حذف نشاط'),
+              content: Text('هل تريد حذف النشاط؟'),
+              actions: <Widget>[
+                CupertinoDialogAction(
+                  child: Text(
+                    'نعم',
+                    style: TextStyle(
+                        color: Colors.red, fontWeight: FontWeight.bold),
+                  ),
+                  onPressed: () {
+                    Provider.of<Activities>(context, listen: false)
+                        .deleteActivity(widget.id, orgId);
+                    Navigator.of(ctx).pop();
+                  },
+                ),
+                CupertinoDialogAction(
+                  child: Text('الغاء'),
+                  onPressed: () {
+                    Navigator.of(ctx).pop();
+                  },
+                ),
+              ],
+            ),
     );
   }
 }
